[New] Created test cases for pages-admin-agency-form component from enzyme to rtl
